Remove the PolyDatabase type
