+++ heed/src/database.rs
@@ -1,15 +1,16 @@
 use std::borrow::Cow;
 use std::ops::{Bound, RangeBounds};
-use std::{fmt, mem, ptr};
+use std::{any, fmt, marker, mem, ptr};
+
+use types::DecodeIgnore;
 
 use crate::mdb::error::mdb_result;
 use crate::mdb::ffi;
-use crate::types::DecodeIgnore;
 use crate::*;
 
-/// A polymorphic database that accepts types on call methods and not at creation.
+/// A typed database that accepts only the types it was created with.
 ///
-/// # Example: Iterate over ranges of databases entries
+/// # Example: Iterate over databases entries
 ///
 /// In this example we store numbers in big endian this way those are ordered.
 /// Thanks to their bytes representation, heed is able to iterate over them
@@ -19,7 +20,7 @@ use crate::*;
 /// # use std::fs;
 /// # use std::path::Path;
 /// # use heed::EnvOpenOptions;
-/// use heed::PolyDatabase;
+/// use heed::Database;
 /// use heed::types::*;
 /// use heed::byteorder::BigEndian;
 ///
@@ -32,29 +33,33 @@ use crate::*;
 /// type BEI64 = I64<BigEndian>;
 ///
 /// let mut wtxn = env.write_txn()?;
-/// let db: PolyDatabase = env.create_poly_database(&mut wtxn, Some("big-endian-iter"))?;
+/// let db: Database<BEI64, Unit> = env.create_database(&mut wtxn, Some("big-endian-iter"))?;
 ///
 /// # db.clear(&mut wtxn)?;
-/// db.put::<BEI64, Unit>(&mut wtxn, &0, &())?;
-/// db.put::<BEI64, Str>(&mut wtxn, &35, "thirty five")?;
-/// db.put::<BEI64, Str>(&mut wtxn, &42, "forty two")?;
-/// db.put::<BEI64, Unit>(&mut wtxn, &68, &())?;
+/// db.put(&mut wtxn, &68, &())?;
+/// db.put(&mut wtxn, &35, &())?;
+/// db.put(&mut wtxn, &0, &())?;
+/// db.put(&mut wtxn, &42, &())?;
 ///
 /// // you can iterate over database entries in order
-/// let range = 35..=42;
-/// let mut range = db.range::<BEI64, Str, _>(&wtxn, &range)?;
-/// assert_eq!(range.next().transpose()?, Some((35, "thirty five")));
-/// assert_eq!(range.next().transpose()?, Some((42, "forty two")));
-/// assert_eq!(range.next().transpose()?, None);
+/// let rets: Result<_, _> = db.iter(&wtxn)?.collect();
+/// let rets: Vec<(i64, _)> = rets?;
 ///
-/// drop(range);
+/// let expected = vec![
+///     (0, ()),
+///     (35, ()),
+///     (42, ()),
+///     (68, ()),
+/// ];
+///
+/// assert_eq!(rets, expected);
 /// wtxn.commit()?;
 /// # Ok(()) }
 /// ```
 ///
-/// # Example: Select ranges of entries
+/// # Example: Iterate over and delete ranges of entries
 ///
-/// Heed also support ranges deletions.
+/// Discern also support ranges and ranges deletions.
 /// Same configuration as above, numbers are ordered, therefore it is safe to specify
 /// a range and be able to iterate over and/or delete it.
 ///
@@ -62,7 +67,7 @@ use crate::*;
 /// # use std::fs;
 /// # use std::path::Path;
 /// # use heed::EnvOpenOptions;
-/// use heed::PolyDatabase;
+/// use heed::Database;
 /// use heed::types::*;
 /// use heed::byteorder::BigEndian;
 ///
@@ -75,20 +80,31 @@ use crate::*;
 /// type BEI64 = I64<BigEndian>;
 ///
 /// let mut wtxn = env.write_txn()?;
-/// let db: PolyDatabase = env.create_poly_database(&mut wtxn, Some("big-endian-iter"))?;
+/// let db: Database<BEI64, Unit> = env.create_database(&mut wtxn, Some("big-endian-iter"))?;
 ///
 /// # db.clear(&mut wtxn)?;
-/// db.put::<BEI64, Unit>(&mut wtxn, &0, &())?;
-/// db.put::<BEI64, Str>(&mut wtxn, &35, "thirty five")?;
-/// db.put::<BEI64, Str>(&mut wtxn, &42, "forty two")?;
-/// db.put::<BEI64, Unit>(&mut wtxn, &68, &())?;
+/// db.put(&mut wtxn, &0, &())?;
+/// db.put(&mut wtxn, &68, &())?;
+/// db.put(&mut wtxn, &35, &())?;
+/// db.put(&mut wtxn, &42, &())?;
+///
+/// // you can iterate over ranges too!!!
+/// let range = 35..=42;
+/// let rets: Result<_, _> = db.range(&wtxn, &range)?.collect();
+/// let rets: Vec<(i64, _)> = rets?;
+///
+/// let expected = vec![
+///     (35, ()),
+///     (42, ()),
+/// ];
+///
+/// assert_eq!(rets, expected);
 ///
 /// // even delete a range of keys
 /// let range = 35..=42;
-/// let deleted = db.delete_range::<BEI64, _>(&mut wtxn, &range)?;
-/// assert_eq!(deleted, 2);
+/// let deleted: usize = db.delete_range(&mut wtxn, &range)?;
 ///
-/// let rets: Result<_, _> = db.iter::<BEI64, Unit>(&wtxn)?.collect();
+/// let rets: Result<_, _> = db.iter(&wtxn)?.collect();
 /// let rets: Vec<(i64, _)> = rets?;
 ///
 /// let expected = vec![
@@ -102,15 +118,15 @@ use crate::*;
 /// wtxn.commit()?;
 /// # Ok(()) }
 /// ```
-#[derive(Copy, Clone)]
-pub struct PolyDatabase {
+pub struct Database<KC, DC> {
     pub(crate) env_ident: usize,
     pub(crate) dbi: ffi::MDB_dbi,
+    marker: marker::PhantomData<(KC, DC)>,
 }
 
-impl PolyDatabase {
-    pub(crate) fn new(env_ident: usize, dbi: ffi::MDB_dbi) -> PolyDatabase {
-        PolyDatabase { env_ident, dbi }
+impl<KC, DC> Database<KC, DC> {
+    pub(crate) fn new(env_ident: usize, dbi: ffi::MDB_dbi) -> Database<KC, DC> {
+        Database { env_ident, dbi, marker: std::marker::PhantomData }
     }
 
     /// Retrieves the value associated with a key.
@@ -131,29 +147,25 @@ impl PolyDatabase {
     /// #     .map_size(10 * 1024 * 1024) // 10MB
     /// #     .max_dbs(3000)
     /// #     .open(dir.path())?;
-    /// type BEI32 = I32<BigEndian>;
+    /// type BEI32= U32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("get-poly-i32"))?;
+    /// let db: Database<Str, BEI32> = env.create_database(&mut wtxn, Some("get-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-two", &42)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-seven", &27)?;
+    /// db.put(&mut wtxn, "i-am-forty-two", &42)?;
+    /// db.put(&mut wtxn, "i-am-twenty-seven", &27)?;
     ///
-    /// let ret = db.get::<Str, BEI32>(&wtxn, "i-am-forty-two")?;
+    /// let ret = db.get(&wtxn, "i-am-forty-two")?;
     /// assert_eq!(ret, Some(42));
     ///
-    /// let ret = db.get::<Str, BEI32>(&wtxn, "i-am-twenty-one")?;
+    /// let ret = db.get(&wtxn, "i-am-twenty-one")?;
     /// assert_eq!(ret, None);
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn get<'a, 'txn, KC, DC>(
-        &self,
-        txn: &'txn RoTxn,
-        key: &'a KC::EItem,
-    ) -> Result<Option<DC::DItem>>
+    pub fn get<'a, 'txn>(&self, txn: &'txn RoTxn, key: &'a KC::EItem) -> Result<Option<DC::DItem>>
     where
         KC: BytesEncode<'a>,
         DC: BytesDecode<'txn>,
@@ -204,26 +216,26 @@ impl PolyDatabase {
     /// type BEU32 = U32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("get-lt-u32"))?;
+    /// let db = env.create_database::<BEU32, Unit>(&mut wtxn, Some("get-lt-u32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &27, &())?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &42, &())?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &43, &())?;
+    /// db.put(&mut wtxn, &27, &())?;
+    /// db.put(&mut wtxn, &42, &())?;
+    /// db.put(&mut wtxn, &43, &())?;
     ///
-    /// let ret = db.get_lower_than::<BEU32, Unit>(&wtxn, &4404)?;
+    /// let ret = db.get_lower_than(&wtxn, &4404)?;
     /// assert_eq!(ret, Some((43, ())));
     ///
-    /// let ret = db.get_lower_than::<BEU32, Unit>(&wtxn, &43)?;
+    /// let ret = db.get_lower_than(&wtxn, &43)?;
     /// assert_eq!(ret, Some((42, ())));
     ///
-    /// let ret = db.get_lower_than::<BEU32, Unit>(&wtxn, &27)?;
+    /// let ret = db.get_lower_than(&wtxn, &27)?;
     /// assert_eq!(ret, None);
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn get_lower_than<'a, 'txn, KC, DC>(
+    pub fn get_lower_than<'a, 'txn>(
         &self,
         txn: &'txn RoTxn,
         key: &'a KC::EItem,
@@ -248,7 +260,7 @@ impl PolyDatabase {
         }
     }
 
-    /// Retrieves the key/value pair lower than or equal the given one in this database.
+    /// Retrieves the key/value pair lower than or equal to the given one in this database.
     ///
     /// If the database if empty or there is no key lower than or equal to the given one,
     /// then `None` is returned.
@@ -272,26 +284,26 @@ impl PolyDatabase {
     /// type BEU32 = U32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("get-lte-u32"))?;
+    /// let db = env.create_database::<BEU32, Unit>(&mut wtxn, Some("get-lt-u32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &27, &())?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &42, &())?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &43, &())?;
+    /// db.put(&mut wtxn, &27, &())?;
+    /// db.put(&mut wtxn, &42, &())?;
+    /// db.put(&mut wtxn, &43, &())?;
     ///
-    /// let ret = db.get_lower_than_or_equal_to::<BEU32, Unit>(&wtxn, &4404)?;
+    /// let ret = db.get_lower_than_or_equal_to(&wtxn, &4404)?;
     /// assert_eq!(ret, Some((43, ())));
     ///
-    /// let ret = db.get_lower_than_or_equal_to::<BEU32, Unit>(&wtxn, &43)?;
+    /// let ret = db.get_lower_than_or_equal_to(&wtxn, &43)?;
     /// assert_eq!(ret, Some((43, ())));
     ///
-    /// let ret = db.get_lower_than_or_equal_to::<BEU32, Unit>(&wtxn, &26)?;
+    /// let ret = db.get_lower_than_or_equal_to(&wtxn, &26)?;
     /// assert_eq!(ret, None);
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn get_lower_than_or_equal_to<'a, 'txn, KC, DC>(
+    pub fn get_lower_than_or_equal_to<'a, 'txn>(
         &self,
         txn: &'txn RoTxn,
         key: &'a KC::EItem,
@@ -344,26 +356,26 @@ impl PolyDatabase {
     /// type BEU32 = U32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("get-lt-u32"))?;
+    /// let db = env.create_database::<BEU32, Unit>(&mut wtxn, Some("get-lt-u32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &27, &())?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &42, &())?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &43, &())?;
+    /// db.put(&mut wtxn, &27, &())?;
+    /// db.put(&mut wtxn, &42, &())?;
+    /// db.put(&mut wtxn, &43, &())?;
     ///
-    /// let ret = db.get_greater_than::<BEU32, Unit>(&wtxn, &0)?;
+    /// let ret = db.get_greater_than(&wtxn, &0)?;
     /// assert_eq!(ret, Some((27, ())));
     ///
-    /// let ret = db.get_greater_than::<BEU32, Unit>(&wtxn, &42)?;
+    /// let ret = db.get_greater_than(&wtxn, &42)?;
     /// assert_eq!(ret, Some((43, ())));
     ///
-    /// let ret = db.get_greater_than::<BEU32, Unit>(&wtxn, &43)?;
+    /// let ret = db.get_greater_than(&wtxn, &43)?;
     /// assert_eq!(ret, None);
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn get_greater_than<'a, 'txn, KC, DC>(
+    pub fn get_greater_than<'a, 'txn>(
         &self,
         txn: &'txn RoTxn,
         key: &'a KC::EItem,
@@ -391,7 +403,7 @@ impl PolyDatabase {
         }
     }
 
-    /// Retrieves the key/value pair greater than or equal the given one in this database.
+    /// Retrieves the key/value pair greater than or equal to the given one in this database.
     ///
     /// If the database if empty or there is no key greater than or equal to the given one,
     /// then `None` is returned.
@@ -415,26 +427,26 @@ impl PolyDatabase {
     /// type BEU32 = U32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("get-lt-u32"))?;
+    /// let db = env.create_database::<BEU32, Unit>(&mut wtxn, Some("get-lt-u32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &27, &())?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &42, &())?;
-    /// db.put::<BEU32, Unit>(&mut wtxn, &43, &())?;
+    /// db.put(&mut wtxn, &27, &())?;
+    /// db.put(&mut wtxn, &42, &())?;
+    /// db.put(&mut wtxn, &43, &())?;
     ///
-    /// let ret = db.get_greater_than_or_equal_to::<BEU32, Unit>(&wtxn, &0)?;
+    /// let ret = db.get_greater_than_or_equal_to(&wtxn, &0)?;
     /// assert_eq!(ret, Some((27, ())));
     ///
-    /// let ret = db.get_greater_than_or_equal_to::<BEU32, Unit>(&wtxn, &42)?;
+    /// let ret = db.get_greater_than_or_equal_to(&wtxn, &42)?;
     /// assert_eq!(ret, Some((42, ())));
     ///
-    /// let ret = db.get_greater_than_or_equal_to::<BEU32, Unit>(&wtxn, &44)?;
+    /// let ret = db.get_greater_than_or_equal_to(&wtxn, &44)?;
     /// assert_eq!(ret, None);
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn get_greater_than_or_equal_to<'a, 'txn, KC, DC>(
+    pub fn get_greater_than_or_equal_to<'a, 'txn>(
         &self,
         txn: &'txn RoTxn,
         key: &'a KC::EItem,
@@ -480,19 +492,19 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("first-poly-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("first-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
     ///
-    /// let ret = db.first::<BEI32, Str>(&wtxn)?;
+    /// let ret = db.first(&wtxn)?;
     /// assert_eq!(ret, Some((27, "i-am-twenty-seven")));
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn first<'txn, KC, DC>(&self, txn: &'txn RoTxn) -> Result<Option<(KC::DItem, DC::DItem)>>
+    pub fn first<'txn>(&self, txn: &'txn RoTxn) -> Result<Option<(KC::DItem, DC::DItem)>>
     where
         KC: BytesDecode<'txn>,
         DC: BytesDecode<'txn>,
@@ -533,19 +545,19 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("last-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("last-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
     ///
-    /// let ret = db.last::<BEI32, Str>(&wtxn)?;
+    /// let ret = db.last(&wtxn)?;
     /// assert_eq!(ret, Some((42, "i-am-forty-two")));
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn last<'txn, KC, DC>(&self, txn: &'txn RoTxn) -> Result<Option<(KC::DItem, DC::DItem)>>
+    pub fn last<'txn>(&self, txn: &'txn RoTxn) -> Result<Option<(KC::DItem, DC::DItem)>>
     where
         KC: BytesDecode<'txn>,
         DC: BytesDecode<'txn>,
@@ -582,24 +594,23 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
     /// let ret = db.len(&wtxn)?;
     /// assert_eq!(ret, 4);
     ///
-    /// db.delete::<BEI32>(&mut wtxn, &27)?;
+    /// db.delete(&mut wtxn, &27)?;
     ///
     /// let ret = db.len(&wtxn)?;
     /// assert_eq!(ret, 3);
     ///
     /// wtxn.commit()?;
-    ///
     /// # Ok(()) }
     /// ```
     pub fn len(&self, txn: &RoTxn) -> Result<u64> {
@@ -636,13 +647,13 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
     /// let ret = db.is_empty(&wtxn)?;
     /// assert_eq!(ret, false);
@@ -684,14 +695,14 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
     ///
-    /// let mut iter = db.iter::<BEI32, Str>(&wtxn)?;
+    /// let mut iter = db.iter(&wtxn)?;
     /// assert_eq!(iter.next().transpose()?, Some((13, "i-am-thirteen")));
     /// assert_eq!(iter.next().transpose()?, Some((27, "i-am-twenty-seven")));
     /// assert_eq!(iter.next().transpose()?, Some((42, "i-am-forty-two")));
@@ -701,7 +712,7 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn iter<'txn, KC, DC>(&self, txn: &'txn RoTxn) -> Result<RoIter<'txn, KC, DC>> {
+    pub fn iter<'txn>(&self, txn: &'txn RoTxn) -> Result<RoIter<'txn, KC, DC>> {
         assert_eq_env_db_txn!(self, txn);
 
         RoCursor::new(txn, self.dbi).map(|cursor| RoIter::new(cursor))
@@ -726,14 +737,14 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
     ///
-    /// let mut iter = db.iter_mut::<BEI32, Str>(&mut wtxn)?;
+    /// let mut iter = db.iter_mut(&mut wtxn)?;
     /// assert_eq!(iter.next().transpose()?, Some((13, "i-am-thirteen")));
     /// let ret = unsafe { iter.del_current()? };
     /// assert!(ret);
@@ -747,22 +758,22 @@ impl PolyDatabase {
     ///
     /// drop(iter);
     ///
-    /// let ret = db.get::<BEI32, Str>(&wtxn, &13)?;
+    /// let ret = db.get(&wtxn, &13)?;
     /// assert_eq!(ret, None);
     ///
-    /// let ret = db.get::<BEI32, Str>(&wtxn, &42)?;
+    /// let ret = db.get(&wtxn, &42)?;
     /// assert_eq!(ret, Some("i-am-the-new-forty-two"));
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn iter_mut<'txn, KC, DC>(&self, txn: &'txn mut RwTxn) -> Result<RwIter<'txn, KC, DC>> {
+    pub fn iter_mut<'txn>(&self, txn: &'txn mut RwTxn) -> Result<RwIter<'txn, KC, DC>> {
         assert_eq_env_db_txn!(self, txn);
 
         RwCursor::new(txn, self.dbi).map(|cursor| RwIter::new(cursor))
     }
 
-    /// Returns a reversed lexicographically ordered iterator of all key-value pairs in this database.
+    /// Return a reversed lexicographically ordered iterator of all key-value pairs in this database.
     ///
     /// ```
     /// # use std::fs;
@@ -781,14 +792,14 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
     ///
-    /// let mut iter = db.rev_iter::<BEI32, Str>(&wtxn)?;
+    /// let mut iter = db.rev_iter(&wtxn)?;
     /// assert_eq!(iter.next().transpose()?, Some((42, "i-am-forty-two")));
     /// assert_eq!(iter.next().transpose()?, Some((27, "i-am-twenty-seven")));
     /// assert_eq!(iter.next().transpose()?, Some((13, "i-am-thirteen")));
@@ -798,14 +809,14 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn rev_iter<'txn, KC, DC>(&self, txn: &'txn RoTxn) -> Result<RoRevIter<'txn, KC, DC>> {
+    pub fn rev_iter<'txn>(&self, txn: &'txn RoTxn) -> Result<RoRevIter<'txn, KC, DC>> {
         assert_eq_env_db_txn!(self, txn);
 
         RoCursor::new(txn, self.dbi).map(|cursor| RoRevIter::new(cursor))
     }
 
-    /// Return a mutable reversed lexicographically ordered iterator of all key-value pairs
-    /// in this database.
+    /// Return a mutable reversed lexicographically ordered iterator of all key-value\
+    /// pairs in this database.
     ///
     /// ```
     /// # use std::fs;
@@ -824,14 +835,14 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
     ///
-    /// let mut iter = db.rev_iter_mut::<BEI32, Str>(&mut wtxn)?;
+    /// let mut iter = db.rev_iter_mut(&mut wtxn)?;
     /// assert_eq!(iter.next().transpose()?, Some((42, "i-am-forty-two")));
     /// let ret = unsafe { iter.del_current()? };
     /// assert!(ret);
@@ -845,19 +856,16 @@ impl PolyDatabase {
     ///
     /// drop(iter);
     ///
-    /// let ret = db.get::<BEI32, Str>(&wtxn, &42)?;
+    /// let ret = db.get(&wtxn, &42)?;
     /// assert_eq!(ret, None);
     ///
-    /// let ret = db.get::<BEI32, Str>(&wtxn, &13)?;
+    /// let ret = db.get(&wtxn, &13)?;
     /// assert_eq!(ret, Some("i-am-the-new-thirteen"));
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn rev_iter_mut<'txn, KC, DC>(
-        &self,
-        txn: &'txn mut RwTxn,
-    ) -> Result<RwRevIter<'txn, KC, DC>> {
+    pub fn rev_iter_mut<'txn>(&self, txn: &'txn mut RwTxn) -> Result<RwRevIter<'txn, KC, DC>> {
         assert_eq_env_db_txn!(self, txn);
 
         RwCursor::new(txn, self.dbi).map(|cursor| RwRevIter::new(cursor))
@@ -884,16 +892,16 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
     /// let range = 27..=42;
-    /// let mut iter = db.range::<BEI32, Str, _>(&wtxn, &range)?;
+    /// let mut iter = db.range(&wtxn, &range)?;
     /// assert_eq!(iter.next().transpose()?, Some((27, "i-am-twenty-seven")));
     /// assert_eq!(iter.next().transpose()?, Some((42, "i-am-forty-two")));
     /// assert_eq!(iter.next().transpose()?, None);
@@ -902,7 +910,7 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn range<'a, 'txn, KC, DC, R>(
+    pub fn range<'a, 'txn, R>(
         &self,
         txn: &'txn RoTxn,
         range: &'a R,
@@ -962,16 +970,16 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
     /// let range = 27..=42;
-    /// let mut range = db.range_mut::<BEI32, Str, _>(&mut wtxn, &range)?;
+    /// let mut range = db.range_mut(&mut wtxn, &range)?;
     /// assert_eq!(range.next().transpose()?, Some((27, "i-am-twenty-seven")));
     /// let ret = unsafe { range.del_current()? };
     /// assert!(ret);
@@ -983,7 +991,7 @@ impl PolyDatabase {
     /// drop(range);
     ///
     ///
-    /// let mut iter = db.iter::<BEI32, Str>(&wtxn)?;
+    /// let mut iter = db.iter(&wtxn)?;
     /// assert_eq!(iter.next().transpose()?, Some((13, "i-am-thirteen")));
     /// assert_eq!(iter.next().transpose()?, Some((42, "i-am-the-new-forty-two")));
     /// assert_eq!(iter.next().transpose()?, Some((521, "i-am-five-hundred-and-twenty-one")));
@@ -993,7 +1001,7 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn range_mut<'a, 'txn, KC, DC, R>(
+    pub fn range_mut<'a, 'txn, R>(
         &self,
         txn: &'txn mut RwTxn,
         range: &'a R,
@@ -1053,16 +1061,16 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
     /// let range = 27..=43;
-    /// let mut iter = db.rev_range::<BEI32, Str, _>(&wtxn, &range)?;
+    /// let mut iter = db.rev_range(&wtxn, &range)?;
     /// assert_eq!(iter.next().transpose()?, Some((42, "i-am-forty-two")));
     /// assert_eq!(iter.next().transpose()?, Some((27, "i-am-twenty-seven")));
     /// assert_eq!(iter.next().transpose()?, None);
@@ -1071,7 +1079,7 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn rev_range<'a, 'txn, KC, DC, R>(
+    pub fn rev_range<'a, 'txn, R>(
         &self,
         txn: &'txn RoTxn,
         range: &'a R,
@@ -1131,16 +1139,16 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
     /// let range = 27..=42;
-    /// let mut range = db.rev_range_mut::<BEI32, Str, _>(&mut wtxn, &range)?;
+    /// let mut range = db.rev_range_mut(&mut wtxn, &range)?;
     /// assert_eq!(range.next().transpose()?, Some((42, "i-am-forty-two")));
     /// let ret = unsafe { range.del_current()? };
     /// assert!(ret);
@@ -1152,7 +1160,7 @@ impl PolyDatabase {
     /// drop(range);
     ///
     ///
-    /// let mut iter = db.iter::<BEI32, Str>(&wtxn)?;
+    /// let mut iter = db.iter(&wtxn)?;
     /// assert_eq!(iter.next().transpose()?, Some((13, "i-am-thirteen")));
     /// assert_eq!(iter.next().transpose()?, Some((27, "i-am-the-new-twenty-seven")));
     /// assert_eq!(iter.next().transpose()?, Some((521, "i-am-five-hundred-and-twenty-one")));
@@ -1162,7 +1170,7 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn rev_range_mut<'a, 'txn, KC, DC, R>(
+    pub fn rev_range_mut<'a, 'txn, R>(
         &self,
         txn: &'txn mut RwTxn,
         range: &'a R,
@@ -1222,16 +1230,16 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<Str, BEI32> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-eight", &28)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-seven", &27)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-nine",  &29)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-one",    &41)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-two",    &42)?;
+    /// db.put(&mut wtxn, "i-am-twenty-eight", &28)?;
+    /// db.put(&mut wtxn, "i-am-twenty-seven", &27)?;
+    /// db.put(&mut wtxn, "i-am-twenty-nine",  &29)?;
+    /// db.put(&mut wtxn, "i-am-forty-one",    &41)?;
+    /// db.put(&mut wtxn, "i-am-forty-two",    &42)?;
     ///
-    /// let mut iter = db.prefix_iter::<Str, BEI32>(&mut wtxn, "i-am-twenty")?;
+    /// let mut iter = db.prefix_iter(&mut wtxn, "i-am-twenty")?;
     /// assert_eq!(iter.next().transpose()?, Some(("i-am-twenty-eight", 28)));
     /// assert_eq!(iter.next().transpose()?, Some(("i-am-twenty-nine", 29)));
     /// assert_eq!(iter.next().transpose()?, Some(("i-am-twenty-seven", 27)));
@@ -1241,7 +1249,7 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn prefix_iter<'a, 'txn, KC, DC>(
+    pub fn prefix_iter<'a, 'txn>(
         &self,
         txn: &'txn RoTxn,
         prefix: &'a KC::EItem,
@@ -1278,16 +1286,16 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<Str, BEI32> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-eight", &28)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-seven", &27)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-nine",  &29)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-one",    &41)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-two",    &42)?;
+    /// db.put(&mut wtxn, "i-am-twenty-eight", &28)?;
+    /// db.put(&mut wtxn, "i-am-twenty-seven", &27)?;
+    /// db.put(&mut wtxn, "i-am-twenty-nine",  &29)?;
+    /// db.put(&mut wtxn, "i-am-forty-one",    &41)?;
+    /// db.put(&mut wtxn, "i-am-forty-two",    &42)?;
     ///
-    /// let mut iter = db.prefix_iter_mut::<Str, BEI32>(&mut wtxn, "i-am-twenty")?;
+    /// let mut iter = db.prefix_iter_mut(&mut wtxn, "i-am-twenty")?;
     /// assert_eq!(iter.next().transpose()?, Some(("i-am-twenty-eight", 28)));
     /// let ret = unsafe { iter.del_current()? };
     /// assert!(ret);
@@ -1301,16 +1309,16 @@ impl PolyDatabase {
     ///
     /// drop(iter);
     ///
-    /// let ret = db.get::<Str, BEI32>(&wtxn, "i-am-twenty-eight")?;
+    /// let ret = db.get(&wtxn, "i-am-twenty-eight")?;
     /// assert_eq!(ret, None);
     ///
-    /// let ret = db.get::<Str, BEI32>(&wtxn, "i-am-twenty-seven")?;
+    /// let ret = db.get(&wtxn, "i-am-twenty-seven")?;
     /// assert_eq!(ret, Some(27000));
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn prefix_iter_mut<'a, 'txn, KC, DC>(
+    pub fn prefix_iter_mut<'a, 'txn>(
         &self,
         txn: &'txn mut RwTxn,
         prefix: &'a KC::EItem,
@@ -1347,16 +1355,16 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<Str, BEI32> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-eight", &28)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-seven", &27)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-nine",  &29)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-one",    &41)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-two",    &42)?;
+    /// db.put(&mut wtxn, "i-am-twenty-eight", &28)?;
+    /// db.put(&mut wtxn, "i-am-twenty-seven", &27)?;
+    /// db.put(&mut wtxn, "i-am-twenty-nine",  &29)?;
+    /// db.put(&mut wtxn, "i-am-forty-one",    &41)?;
+    /// db.put(&mut wtxn, "i-am-forty-two",    &42)?;
     ///
-    /// let mut iter = db.rev_prefix_iter::<Str, BEI32>(&mut wtxn, "i-am-twenty")?;
+    /// let mut iter = db.rev_prefix_iter(&mut wtxn, "i-am-twenty")?;
     /// assert_eq!(iter.next().transpose()?, Some(("i-am-twenty-seven", 27)));
     /// assert_eq!(iter.next().transpose()?, Some(("i-am-twenty-nine", 29)));
     /// assert_eq!(iter.next().transpose()?, Some(("i-am-twenty-eight", 28)));
@@ -1366,7 +1374,7 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn rev_prefix_iter<'a, 'txn, KC, DC>(
+    pub fn rev_prefix_iter<'a, 'txn>(
         &self,
         txn: &'txn RoTxn,
         prefix: &'a KC::EItem,
@@ -1381,7 +1389,7 @@ impl PolyDatabase {
         RoCursor::new(txn, self.dbi).map(|cursor| RoRevPrefix::new(cursor, prefix_bytes))
     }
 
-    /// Return a mutable lexicographically ordered iterator of all key-value pairs
+    /// Return a mutable reversed lexicographically ordered iterator of all key-value pairs
     /// in this database that starts with the given prefix.
     ///
     /// Comparisons are made by using the bytes representation of the key.
@@ -1403,16 +1411,16 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<Str, BEI32> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-eight", &28)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-seven", &27)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-twenty-nine",  &29)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-one",    &41)?;
-    /// db.put::<Str, BEI32>(&mut wtxn, "i-am-forty-two",    &42)?;
+    /// db.put(&mut wtxn, "i-am-twenty-eight", &28)?;
+    /// db.put(&mut wtxn, "i-am-twenty-seven", &27)?;
+    /// db.put(&mut wtxn, "i-am-twenty-nine",  &29)?;
+    /// db.put(&mut wtxn, "i-am-forty-one",    &41)?;
+    /// db.put(&mut wtxn, "i-am-forty-two",    &42)?;
     ///
-    /// let mut iter = db.rev_prefix_iter_mut::<Str, BEI32>(&mut wtxn, "i-am-twenty")?;
+    /// let mut iter = db.rev_prefix_iter_mut(&mut wtxn, "i-am-twenty")?;
     /// assert_eq!(iter.next().transpose()?, Some(("i-am-twenty-seven", 27)));
     /// let ret = unsafe { iter.del_current()? };
     /// assert!(ret);
@@ -1426,16 +1434,16 @@ impl PolyDatabase {
     ///
     /// drop(iter);
     ///
-    /// let ret = db.get::<Str, BEI32>(&wtxn, "i-am-twenty-seven")?;
+    /// let ret = db.get(&wtxn, "i-am-twenty-seven")?;
     /// assert_eq!(ret, None);
     ///
-    /// let ret = db.get::<Str, BEI32>(&wtxn, "i-am-twenty-eight")?;
+    /// let ret = db.get(&wtxn, "i-am-twenty-eight")?;
     /// assert_eq!(ret, Some(28000));
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn rev_prefix_iter_mut<'a, 'txn, KC, DC>(
+    pub fn rev_prefix_iter_mut<'a, 'txn>(
         &self,
         txn: &'txn mut RwTxn,
         prefix: &'a KC::EItem,
@@ -1450,7 +1458,7 @@ impl PolyDatabase {
         RwCursor::new(txn, self.dbi).map(|cursor| RwRevPrefix::new(cursor, prefix_bytes))
     }
 
-    /// Insert a key-value pair in this database.
+    /// Insert a key-value pairs in this database.
     ///
     /// ```
     /// # use std::fs;
@@ -1469,26 +1477,21 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
-    /// let ret = db.get::<BEI32, Str>(&mut wtxn, &27)?;
+    /// let ret = db.get(&mut wtxn, &27)?;
     /// assert_eq!(ret, Some("i-am-twenty-seven"));
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn put<'a, KC, DC>(
-        &self,
-        txn: &mut RwTxn,
-        key: &'a KC::EItem,
-        data: &'a DC::EItem,
-    ) -> Result<()>
+    pub fn put<'a>(&self, txn: &mut RwTxn, key: &'a KC::EItem, data: &'a DC::EItem) -> Result<()>
     where
         KC: BytesEncode<'a>,
         DC: BytesEncode<'a>,
@@ -1529,21 +1532,21 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db = env.create_database::<BEI32, Str>(&mut wtxn, Some("number-string"))?;
     ///
     /// # db.clear(&mut wtxn)?;
     /// let value = "I am a long long long value";
-    /// db.put_reserved::<BEI32, _>(&mut wtxn, &42, value.len(), |reserved| {
+    /// db.put_reserved(&mut wtxn, &42, value.len(), |reserved| {
     ///     reserved.write_all(value.as_bytes())
     /// })?;
     ///
-    /// let ret = db.get::<BEI32, Str>(&mut wtxn, &42)?;
+    /// let ret = db.get(&mut wtxn, &42)?;
     /// assert_eq!(ret, Some(value));
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn put_reserved<'a, KC, F>(
+    pub fn put_reserved<'a, F>(
         &self,
         txn: &mut RwTxn,
         key: &'a KC::EItem,
@@ -1596,29 +1599,24 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("append-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.append::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.append::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.append::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.append::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.append(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.append(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.append(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.append(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
-    /// let ret = db.get::<BEI32, Str>(&mut wtxn, &27)?;
+    /// let ret = db.get(&mut wtxn, &27)?;
     /// assert_eq!(ret, Some("i-am-twenty-seven"));
     ///
     /// // Be wary if you insert at the end unsorted you get the KEYEXIST error.
-    /// assert!(db.append::<BEI32, Str>(&mut wtxn, &1, "Oh No").is_err());
+    /// assert!(db.append(&mut wtxn, &1, "Oh No").is_err());
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn append<'a, KC, DC>(
-        &self,
-        txn: &mut RwTxn,
-        key: &'a KC::EItem,
-        data: &'a DC::EItem,
-    ) -> Result<()>
+    pub fn append<'a>(&self, txn: &mut RwTxn, key: &'a KC::EItem, data: &'a DC::EItem) -> Result<()>
     where
         KC: BytesEncode<'a>,
         DC: BytesEncode<'a>,
@@ -1660,27 +1658,27 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
-    /// let ret = db.delete::<BEI32>(&mut wtxn, &27)?;
+    /// let ret = db.delete(&mut wtxn, &27)?;
     /// assert_eq!(ret, true);
     ///
-    /// let ret = db.get::<BEI32, Str>(&mut wtxn, &27)?;
+    /// let ret = db.get(&mut wtxn, &27)?;
     /// assert_eq!(ret, None);
     ///
-    /// let ret = db.delete::<BEI32>(&mut wtxn, &467)?;
+    /// let ret = db.delete(&mut wtxn, &467)?;
     /// assert_eq!(ret, false);
     ///
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn delete<'a, KC>(&self, txn: &mut RwTxn, key: &'a KC::EItem) -> Result<bool>
+    pub fn delete<'a>(&self, txn: &mut RwTxn, key: &'a KC::EItem) -> Result<bool>
     where
         KC: BytesEncode<'a>,
     {
@@ -1726,19 +1724,20 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
     /// let range = 27..=42;
-    /// let ret = db.delete_range::<BEI32, _>(&mut wtxn, &range)?;
+    /// let ret = db.delete_range(&mut wtxn, &range)?;
     /// assert_eq!(ret, 2);
     ///
-    /// let mut iter = db.iter::<BEI32, Str>(&wtxn)?;
+    ///
+    /// let mut iter = db.iter(&wtxn)?;
     /// assert_eq!(iter.next().transpose()?, Some((13, "i-am-thirteen")));
     /// assert_eq!(iter.next().transpose()?, Some((521, "i-am-five-hundred-and-twenty-one")));
     /// assert_eq!(iter.next().transpose()?, None);
@@ -1747,7 +1746,7 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn delete_range<'a, 'txn, KC, R>(&self, txn: &'txn mut RwTxn, range: &'a R) -> Result<usize>
+    pub fn delete_range<'a, 'txn, R>(&self, txn: &'txn mut RwTxn, range: &'a R) -> Result<usize>
     where
         KC: BytesEncode<'a> + BytesDecode<'txn>,
         R: RangeBounds<KC::EItem>,
@@ -1755,7 +1754,7 @@ impl PolyDatabase {
         assert_eq_env_db_txn!(self, txn);
 
         let mut count = 0;
-        let mut iter = self.range_mut::<KC, DecodeIgnore, _>(txn, range)?;
+        let mut iter = self.remap_data_type::<DecodeIgnore>().range_mut(txn, range)?;
 
         while iter.next().is_some() {
             // safety: We do not keep any reference from the database while using `del_current`.
@@ -1792,13 +1791,13 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<BEI32, Str> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &42, "i-am-forty-two")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &27, "i-am-twenty-seven")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &13, "i-am-thirteen")?;
-    /// db.put::<BEI32, Str>(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
+    /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
+    /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
+    /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
+    /// db.put(&mut wtxn, &521, "i-am-five-hundred-and-twenty-one")?;
     ///
     /// db.clear(&mut wtxn)?;
     ///
@@ -1814,7 +1813,7 @@ impl PolyDatabase {
         unsafe { mdb_result(ffi::mdb_drop(txn.txn.txn, self.dbi, 0)).map_err(Into::into) }
     }
 
-    /// Read this polymorphic database like a typed one, specifying the codecs.
+    /// Change the codec types of this uniform database, specifying the codecs.
     ///
     /// # Safety
     ///
@@ -1828,7 +1827,7 @@ impl PolyDatabase {
     /// # use std::fs;
     /// # use std::path::Path;
     /// # use heed::EnvOpenOptions;
-    /// use heed::{Database, PolyDatabase};
+    /// use heed::Database;
     /// use heed::types::*;
     /// use heed::byteorder::BigEndian;
     ///
@@ -1841,11 +1840,11 @@ impl PolyDatabase {
     /// type BEI32 = I32<BigEndian>;
     ///
     /// let mut wtxn = env.write_txn()?;
-    /// let db = env.create_poly_database(&mut wtxn, Some("iter-i32"))?;
+    /// let db: Database<Unit, Unit> = env.create_database(&mut wtxn, Some("iter-i32"))?;
     ///
     /// # db.clear(&mut wtxn)?;
     /// // We remap the types for ease of use.
-    /// let db = db.as_uniform::<BEI32, Str>();
+    /// let db = db.remap_types::<BEI32, Str>();
     /// db.put(&mut wtxn, &42, "i-am-forty-two")?;
     /// db.put(&mut wtxn, &27, "i-am-twenty-seven")?;
     /// db.put(&mut wtxn, &13, "i-am-thirteen")?;
@@ -1854,13 +1853,39 @@ impl PolyDatabase {
     /// wtxn.commit()?;
     /// # Ok(()) }
     /// ```
-    pub fn as_uniform<KC, DC>(&self) -> Database<KC, DC> {
+    pub fn remap_types<KC2, DC2>(&self) -> Database<KC2, DC2> {
         Database::new(self.env_ident, self.dbi)
+    }
+
+    /// Change the key codec type of this uniform database, specifying the new codec.
+    pub fn remap_key_type<KC2>(&self) -> Database<KC2, DC> {
+        self.remap_types::<KC2, DC>()
+    }
+
+    /// Change the data codec type of this uniform database, specifying the new codec.
+    pub fn remap_data_type<DC2>(&self) -> Database<KC, DC2> {
+        self.remap_types::<KC, DC2>()
+    }
+
+    /// Wrap the data bytes into a lazy decoder.
+    pub fn lazily_decode_data(&self) -> Database<KC, LazyDecode<DC>> {
+        self.remap_types::<KC, LazyDecode<DC>>()
     }
 }
 
-impl fmt::Debug for PolyDatabase {
+impl<KC, DC> Clone for Database<KC, DC> {
+    fn clone(&self) -> Database<KC, DC> {
+        Database { env_ident: self.env_ident, dbi: self.dbi, marker: marker::PhantomData }
+    }
+}
+
+impl<KC, DC> Copy for Database<KC, DC> {}
+
+impl<KC, DC> fmt::Debug for Database<KC, DC> {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.debug_struct("PolyDatabase").finish()
+        f.debug_struct("Database")
+            .field("key_codec", &any::type_name::<KC>())
+            .field("data_codec", &any::type_name::<DC>())
+            .finish()
     }
 }
