--- heed/src/db/mod.rs
@@ -1,5 +1,0 @@
-mod polymorph;
-mod uniform;
-
-pub use self::polymorph::PolyDatabase;
-pub use self::uniform::Database;
