+++ lmdb-master-sys/tests/lmdb.rs
@@ -9,10 +9,8 @@ fn test_lmdb() {
     lmdb.push("libraries");
     lmdb.push("liblmdb");
 
-    let make_cmd = Command::new("make")
-        .current_dir(&lmdb)
-        .status()
-        .expect("failed to execute process");
+    let make_cmd =
+        Command::new("make").current_dir(&lmdb).status().expect("failed to execute process");
 
     assert_eq!(make_cmd.success(), true);
