+++ lmdb-master-sys/tests/simple.rs
@@ -1,9 +1,9 @@
-use cstr::cstr;
-use lmdb_master_sys::*;
-
 use std::ffi::{c_void, CString};
 use std::fs::{self, File};
 use std::ptr;
+
+use cstr::cstr;
+use lmdb_master_sys::*;
 
 // https://github.com/victorporof/lmdb/blob/mdb.master/libraries/liblmdb/moz-test.c
 
@@ -46,14 +46,8 @@ fn test_simple(env_path: &str) {
 
     let mut env: *mut MDB_env = ptr::null_mut();
     let mut dbi: MDB_dbi = 0;
-    let mut key = MDB_val {
-        mv_size: 0,
-        mv_data: ptr::null_mut(),
-    };
-    let mut data = MDB_val {
-        mv_size: 0,
-        mv_data: ptr::null_mut(),
-    };
+    let mut key = MDB_val { mv_size: 0, mv_data: ptr::null_mut() };
+    let mut data = MDB_val { mv_size: 0, mv_data: ptr::null_mut() };
     let mut txn: *mut MDB_txn = ptr::null_mut();
     let sval = cstr!("foo").as_ptr() as *mut c_void;
     let dval = cstr!("bar").as_ptr() as *mut c_void;
@@ -65,12 +59,7 @@ fn test_simple(env_path: &str) {
         E!(mdb_env_open(env, env_path.as_ptr(), 0, 0664));
 
         E!(mdb_txn_begin(env, ptr::null_mut(), 0, &mut txn));
-        E!(mdb_dbi_open(
-            txn,
-            cstr!("subdb").as_ptr(),
-            MDB_CREATE,
-            &mut dbi
-        ));
+        E!(mdb_dbi_open(txn, cstr!("subdb").as_ptr(), MDB_CREATE, &mut dbi));
         E!(mdb_txn_commit(txn));
 
         key.mv_size = 3;
